Unix poll
