+++ Cargo.toml
@@ -15,3 +15,7 @@ edition = "2018"
 [dependencies]
 openssl = "0.10.30"
 rand = "0.7.3"
+
+[build-dependencies]
+cc = { version = "1.0.59", features = ["parallel"] }
+bindgen = "0.55.0"
